requirements.txt
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
lxml==5.1.0
beautifulsoup4==4.12.3
pandas-datareader==0.10.0
fredapi==0.5.1
scipy==1.13.0
scikit-learn==1.4.2
statsmodels==0.14.2
streamlit==1.35.0
plotly==5.22.0
pyarrow==16.0.0
python-dotenv==1.0.1
certifi

# Optional: Google Sheets subscriber backend
# google-auth>=2.29.0
# google-api-python-client>=2.127.0
